# openapi-live.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "s-curve",
    "description": "",
    "contact": {
      "name": "s-curve team"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8800"
    },
    {
      "url": "https://rust-service:8800"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users (admin)",
        "description": "Returns a paginated list of users, optionally filtered by name or email.\nAccessible only to admins or users with `user.view` permission.",
        "operationId": "list_users",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Optional search query (name or email)",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based, default 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page (default 25, max 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List users",
            "headers": {
              "X-Total-Count": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Total number of users matching the query"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "example": [
                  {
                    "created_at": "2025-01-15T10:00:00Z",
                    "email": "ada@example.com",
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "name": "Ada Lovelace",
                    "provider": "local",
                    "updated_at": "2025-01-15T10:00:00Z"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user (admin)",
        "description": "Creates a new user with password hash. Requires `user.manage` permission.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              },
              "example": {
                "email": "dev1@example.com",
                "name": "Developer One",
                "password": "SecurePassword123!"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "created_at": "2025-01-15T10:00:00Z",
                  "email": "ada@example.com",
                  "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "name": "Ada Lovelace",
                  "provider": "local",
                  "updated_at": "2025-01-15T10:00:00Z"
                }
              }
            }
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "description": "Returns a paginated list of users, optionally filtered by name or email.\nAccessible only to admins or users with `user.view` permission.",
        "operationId": "list_users",
        "parameters": [
          {
            "description": "Optional search query (name or email)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Page number (1-based, default 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Items per page (default 25, max 100)",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "created_at": "2025-01-15T10:00:00Z",
                    "email": "ada@example.com",
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "name": "Ada Lovelace",
                    "provider": "local",
                    "updated_at": "2025-01-15T10:00:00Z"
                  }
                ],
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List users",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of users matching the query",
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List users (admin)",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "description": "Creates a new user with password hash. Requires `user.manage` permission.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "dev1@example.com",
                "name": "Developer One",
                "password": "SecurePassword123!"
              },
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2025-01-15T10:00:00Z",
                  "email": "ada@example.com",
                  "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "name": "Ada Lovelace",
                  "provider": "local",
                  "updated_at": "2025-01-15T10:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "User created"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new user (admin)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user (admin)",
        "description": "Updates user details. Requires `user.manage` permission.",
        "operationId": "update_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "example": {
                "email": "dev2@example.com",
                "name": "Developer Two"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "created_at": "2025-01-15T10:00:00Z",
                  "email": "ada@example.com",
                  "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "name": "Ada Lovelace",
                  "provider": "local",
                  "updated_at": "2025-01-15T10:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user (admin)",
        "description": "Soft-deletes a user by setting deleted_at. Requires `user.manage` permission.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "00000000-0000-0000-0000-000000000000"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                },
                "example": {
                  "message": "User deleted"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "description": "Soft-deletes a user by setting deleted_at. Requires `user.manage` permission.",
        "operationId": "delete_user",
        "parameters": [
          {
            "description": "User ID",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "message": "User deleted"
                },
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            },
            "description": "User deleted"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a user (admin)",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "description": "Updates user details. Requires `user.manage` permission.",
        "operationId": "update_user",
        "parameters": [
          {
            "description": "User ID",
            "example": "00000000-0000-0000-0000-000000000000",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "dev2@example.com",
                "name": "Developer Two"
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2025-01-15T10:00:00Z",
                  "email": "ada@example.com",
                  "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "name": "Ada Lovelace",
                  "provider": "local",
                  "updated_at": "2025-01-15T10:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "User updated"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a user (admin)",
        "tags": [
          "Users"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "newadmin@example.com"
          },
          "name": {
            "type": "string",
            "example": "New Admin"
          },
          "password": {
            "type": "string",
            "example": "SecureP@ss123"
          }
        }
      },
...
      "DeletedResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "updated@example.com",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Updated Name",
            "nullable": true
          },
          "password": {
            "type": "string",
            "example": "NewP@ssword456",
            "nullable": true
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "provider",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "provider_id": {
            "type": "string",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}